+++ DataRepo/tests/loaders/test_sequences_loader.py
@@ -1,9 +1,9 @@
 import pandas as pd
 
+from DataRepo.loaders.sequences_loader import SequencesLoader
 from DataRepo.models import LCMethod, MSRunSequence
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 from DataRepo.utils.file_utils import read_from_file
-from DataRepo.utils.sequences_loader import SequencesLoader
 
 
 class SequencesLoaderTests(TracebaseTestCase):
