+++ DataRepo/tests/management/test_lcms_metadata_loading.py
@@ -4,16 +4,19 @@ from datetime import datetime
 import pandas as pd
 from django.core.management import call_command
 
+from DataRepo.loaders.accucor_data_loader import AccuCorDataLoader
+from DataRepo.loaders.sample_table_loader import (
+    LCMSDBSampleMissing,
+    SampleTableLoader,
+)
 from DataRepo.models import LCMethod, MaintainedModel, MSRunSample, Sample
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 from DataRepo.utils import (  # TODO: Uncomment when issue #814 is implemented; NoMZXMLFiles,
-    AccuCorDataLoader,
     AggregatedErrors,
     AggregatedErrorsSet,
     DuplicateSampleDataHeaders,
     InvalidLCMSHeaders,
     LCMethodFixturesMissing,
-    LCMSDBSampleMissing,
     LCMSDefaultsRequired,
     MismatchedSampleHeaderMZXML,
     MissingLCMSSampleDataHeaders,
@@ -21,7 +24,6 @@ from DataRepo.utils import (  # TODO: Uncomment when issue #814 is implemented; 
     MissingPeakAnnotationFiles,
     MissingRequiredLCMSValues,
     PeakAnnotFileMismatches,
-    SampleTableLoader,
     UnexpectedLCMSSampleDataHeaders,
     read_from_file,
 )
