+++ DataRepo/tests/management/test_load_protocols.py
@@ -4,9 +4,9 @@ import pandas as pd
 from django.core.management import call_command
 from django.test import tag
 
+from DataRepo.loaders import ProtocolsLoader
 from DataRepo.models import Protocol
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
-from DataRepo.utils import ProtocolsLoader
 from DataRepo.utils.exceptions import (
     AggregatedErrors,
     ConflictingValueErrors,
